requirements.txt
openai==1.54.3
requests==2.31.0
Pillow==10.4.0
jinja2==3.1.4
colorama==0.4.6
python-dotenv==1.0.1
aiohttp==3.9.5
asyncio
concurrent.futures
urllib3==2.0.7